+++ flask_via/examples/small/views.py
@@ -1,0 +1,20 @@
+# -*- coding: utf-8 -*-
+
+"""
+flask_via.examples.small.views
+==============================
+
+A small ``Flask-Via`` example Flask application.
+"""
+
+
+def home():
+    return 'Home Page'
+
+
+def contact():
+    return 'Contact Page'
+
+
+def about():
+    return 'About Page'