+++ CHANGELOG.rst
@@ -1,5 +1,8 @@
 Change Log
 ==========
 
-Initial Release
----------------
+2014.05.02
+----------
+* Feature: Flask extension initialisation
+* Feature: Basic and Pluggable Flask Routers
+* Feature: Flask-Restful Router